Fixed Header

+++ sixDegreeGraph/sixDegreeGraph/main.cpp
@@ -35,7 +35,7 @@ int main()
 	//First we will turn every follower into a user object. Could do in above user loop.
 
 	Graph userGraph(listOfIds.size());// Insert graph size based on Followers data
-	for (int iter = 0; iter < listOfIds.size(); iter++) {
+	for (size_t iter = 0; iter < listOfIds.size(); iter++) {
 		userGraph.addEdge(0, listOfIds[iter]); // Insert an edge by grabbing the Users ID and connecting it to either the orig user or the target account.
 	}
